move 'reduceWhile' tests to separate file
