+++ types/ramda/ramda-tests.ts
@@ -178,30 +178,6 @@ R.times(i, 5);
 };
 
 () => {
-    const isOdd = (acc: number, x: number) => x % 2 === 1;
-
-    const xs: number[] = [1, 3, 5, 60, 777, 800];
-    R.reduceWhile(isOdd, R.add, 0, xs); // => 9
-    R.reduceWhile(isOdd)(R.add, 0, xs); // => 9
-    R.reduceWhile(isOdd)(R.add, 0)(xs); // => 9
-    R.reduceWhile(isOdd)(R.add)(0, xs); // => 9
-    R.reduceWhile(isOdd)(R.add)(0)(xs); // => 9
-    R.reduceWhile(isOdd, R.add)(0, xs); // => 9
-    R.reduceWhile(isOdd, R.add)(0)(xs); // => 9
-    R.reduceWhile(isOdd, R.add, 0)(xs); // => 9
-
-    const ys: number[] = [];
-    R.reduceWhile(isOdd, R.add, 111, ys); // => 111
-    R.reduceWhile(isOdd)(R.add, 111, ys); // => 111
-    R.reduceWhile(isOdd)(R.add, 111)(ys); // => 111
-    R.reduceWhile(isOdd)(R.add)(111, ys); // => 111
-    R.reduceWhile(isOdd)(R.add)(111)(ys); // => 111
-    R.reduceWhile(isOdd, R.add)(111, ys); // => 111
-    R.reduceWhile(isOdd, R.add)(111)(ys); // => 111
-    R.reduceWhile(isOdd, R.add, 111)(ys); // => 111
-};
-
-() => {
     function isOdd(n: number) {
         return n % 2 === 1;
     }
